// packages/sidetrack/src/models/generated/public/SidetrackJobStatusEnum.ts
// @generated
// This file is automatically generated by Kanel. Do not modify manually.

/** Represents the enum public.sidetrack_job_status_enum */
type SidetrackJobStatusEnum = 
  | 'scheduled'
  | 'running'
  | 'cancelled'
  | 'failed'
  | 'retrying'
  | 'completed';

export default SidetrackJobStatusEnum;